requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15.0
drf-spectacular[sidecar]>=0.27.0
python-decouple>=3.8